Add interfaces for save
We need help from the compiler so these interfaces will be useful for determining the shape of the data.
