+++ src/index.ts
@@ -67,29 +67,9 @@ import {Transaction} from './transaction';
 import {promisifyAll} from '@google-cloud/promisify';
 import {google} from '../protos/protos';
 import {AggregateQuery} from './aggregate';
+import {SaveEntity} from './interfaces/save';
 
 const {grpc} = new GrpcClient();
-
-interface SaveEntityWithKeySymbol {
-  [entity.KEY_SYMBOL]: SaveArrayData[] | SaveNonArrayData;
-}
-
-type SaveNonArrayData = google.datastore.v1.IEntity;
-
-interface SaveArrayData {
-  name: string;
-  value: google.datastore.v1.IValue;
-  excludeFromIndexes: boolean;
-}
-
-interface SaveEntityWithoutKeySymbol {
-  key: entity.Key;
-  data: SaveArrayData[] | SaveNonArrayData;
-}
-
-type SaveEntity = SaveEntityWithoutKeySymbol;
-
-type SaveEntities = SaveEntity | SaveEntity[];
 
 export type PathType = string | number | entity.Int;
 export interface BooleanObject {
@@ -1093,11 +1073,11 @@ class Datastore extends DatastoreRequest {
   ): void;
   save(entities: Entities, callback: SaveCallback): void;
   save(
-    entities: SaveEntities,
+    entities: Entities,
     gaxOptionsOrCallback?: CallOptions | SaveCallback,
     cb?: SaveCallback
   ): void | Promise<SaveResponse> {
-    const arrayifiedEntities = arrify(entities);
+    entities = arrify(entities) as SaveEntity[];
     const gaxOptions =
       typeof gaxOptionsOrCallback === 'object' ? gaxOptionsOrCallback : {};
     const callback =
@@ -1113,7 +1093,7 @@ class Datastore extends DatastoreRequest {
 
     // Iterate over the entity objects, build a proto from all keys and values,
     // then place in the correct mutation array (insert, update, etc).
-    arrayifiedEntities
+    entities
       .map(DatastoreRequest.prepareEntityObject_)
       .forEach((entityObject: Entity, index: number) => {
         const mutation: Mutation = {};
@@ -1205,7 +1185,7 @@ class Datastore extends DatastoreRequest {
         }
         if (insertIndexes[index]) {
           const id = entity.keyFromKeyProto(result.key).id;
-          arrayifiedEntities[index].key.id = id;
+          entities[index].key.id = id;
         }
       });
